Refs #197 - Encapsulate create_vm command
