+++ lib/fog/vsphere/requests/compute/get_cluster.rb
@@ -11,12 +11,7 @@ module Fog
         protected
 
         def get_raw_cluster(name, datacenter_name_or_obj)
-          dc = if datacenter_name_or_obj.is_a?(String)
-                 find_raw_datacenter(datacenter_name_or_obj)
-               else
-                 datacenter_name_or_obj
-               end
-          dc.find_compute_resource(name)
+          shared_request.get_raw_cluster(name, datacenter_name_or_obj)
         end
       end
