+++ lib/fog/vsphere/requests/compute/get_host.rb
@@ -9,9 +9,7 @@ module Fog
         protected
 
         def get_raw_host(name, cluster_name, datacenter_name)
-          cluster = get_raw_cluster(cluster_name, datacenter_name)
-          cluster.host.find { |host| host.name == name } ||
-            raise(Fog::Vsphere::Compute::NotFound, "no such host #{name}")
+          shared_request.get_raw_host(name, cluster_name, datacenter_name)
         end
       end
     end
